DHD: Prevent waiting for file dump event during reboot and reduce bus busy timeout
Problem:
Customer reported that reboot is timing out due to
DHD waiting for file dump event and also due to the
DHD_BUS_BUSY_TIMEOUT being 30s.
There was an ROT prior to reboot, and while DHD was
waiting for the file dump event
from HAL to collect mem dumps, reboot was initiated,
due to which the file dump event does not arrive and
DHD times out after waiting for 30s, which is too
much during reboot time.

Fix:
1. Reduce the timeout to 5s
2. Do not wait for file dump event and return error if reboot is already in progress
3. If file dump event times out during reboot, avoid collecting dumps

Bug: 243437575

Change-Id: Iba5e01cfc9aed028623d1ec56cd4dcbf746bba1b
Signed-off-by: Yadukishore Kodicherla <yadukishore.kodicherla@broadcom.corp-partner.google.com>