DHD: 4398 - do not start rpm timer from dhd_bus_start
Customer complained that in every 10s they are observing nearly 100 wakeups of
dhd_rpm_state thread despite only BT being switched on.

Cause: This is because when BT is switched on, WLAN is also internally
turned on as per ASIC's requirement. As part of this dhd_bus_start is
called where RPM timer is started , which wakes up the rpm thread every 100ms.

Fix: RPM timer is anyway started as part of dhd_open, so remove rpm timer start from dhd_bus_start

Bug: 271952233

Test: hikey 4398d0 with prints in rpm thread

Change-Id: Ibe42ae961b2ebbc4afac617ec71aeed1df407f8c
Signed-off-by: Yadukishore Kodicherla <yadukishore.kodicherla@broadcom.corp-partner.google.com>