bcmdhd: Fix for wrong reporting of link id in the ML connect done event
kernel requires, the link data to be populated as an array starting with index 0. However, DHD was populating the values with link_idx which may not start with 0 and in those cases, the connection indication will fail.

This was supposed to be fixed by below RB and validated, but after resolving merge conflict, one of the line got mistakenly deleted. Putting back the missing line.

Bug: 280514629
Test: BRCM Internal SVT test

Change-Id: I498b2b0cf5ff326127172744c2813e182a69fd59
Signed-off-by: Sungjoon Park <sungjoon.park@broadcom.corp-partner.google.com>